refactor(SW-35): split IsoDeco into sub-components, hook, and constants
Extracts the monolithic component into a proper module structure:
constants/types/hook files, and Caption, Header, Layer, Layers
sub-components. No behaviour change intended.
