+++ src/components/features/HomePage/Hero/components/IsoDeco/IsoDeco.types.ts
@@ -1,0 +1,8 @@
+export type Phase = 'fade' | 'stack' | 'done';
+
+export interface LayerConfig {
+  i: number;
+  spreadY: number;
+  finalY: number;
+  hoverDy: number;
+}
